num atoms and more refactoring
