+++ requirements/pip.txt
@@ -1,2 +1,5 @@
 numpy
 matplotlib
+pandas
+pexpect
+cifkit
